tests + results of integrating with parsers
Tried integrating with parsers library.
But encountered infinite loops.
For instance -  Text.Parser.Token.integer

I suspect this is to do with the way bindLexer was written. We have
built-in backtracking. In order for me to try to lexers, I just need to
use "<|>" of Alternative type class. However, the Parsing typeclass is
expecting an implementation of 'try'. Which is id for me.
Another hint is unless eof was also implemented, it was throwing a
compilation error. Meaning, at the least, there is an impedance
mismatch.

Another problem with the parsers library is  that the tokens are all
based on Haskell Report. For instance, integer is defined
based on the syntax of integer literals in Haskell. It means, no
binaries.
I really want support for multiple Languages. So we are ignoring this
for now.
Only nagging issue is - we still do not know if my approach is going to
hurt me later on.